+++ lib/Frontend/CompilerInvocation.cpp
@@ -1953,22 +1953,6 @@ static bool ParseLangArgs(LangOptions &Opts, ArgList &Args,
     Opts.UseCheckedAsyncObjCBridging = true;
   }
 
-  if (auto A = Args.getLastArg(OPT_objc_call_swift_async_bridging)) {
-    using Mode = LangOptions::ObjcToSwiftAsyncBridgingMode;
-    auto value =
-        llvm::StringSwitch<std::optional<Mode>>(A->getValue())
-            .Case("task", Mode::Task)
-            .Case("immediate", Mode::TaskImmediate)
-            .Default(std::nullopt);
-    if (value) {
-      Opts.objcToSwiftAsyncBridgingMode = *value;
-    } else {
-      Diags.diagnose(SourceLoc(), diag::error_invalid_arg_value,
-                     A->getAsString(Args), A->getValue());
-      HadError = true;
-    }
-  }
-
   Opts.DisableDynamicActorIsolation |=
       Args.hasArg(OPT_disable_dynamic_actor_isolation);